refactor : confing -> config 디렉토리 이름 오타 수정 (#30)
