+++ src/main/java/org/chzz/market/domain/auction/entity/Auction.java
@@ -12,9 +12,7 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import jakarta.validation.constraints.Pattern;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.*;
 import org.chzz.market.domain.base.entity.BaseTimeEntity;
 import org.chzz.market.domain.product.entity.Product;
 import org.chzz.market.domain.user.entity.User;
@@ -22,6 +20,8 @@ import org.chzz.market.domain.user.entity.User;
 @Getter
 @Entity
 @Table
+@Builder
+@AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Auction extends BaseTimeEntity {
     @Id
@@ -38,7 +38,8 @@ public class Auction extends BaseTimeEntity {
     private User winner;
 
     @Column
-    @Pattern(regexp = "^[1-9][0-9]*000$")
+    // TODO: @Pattern은 문자열에서만 적용가능 해서 임시 주석 처리 추후 수정 필요
+    // @Pattern(regexp = "^[1-9][0-9]*000$")
     private Long minPrice;
 
     @Column(columnDefinition = "varchar(20)")
