+++ src/test/java/org/chzz/market/common/DatabaseTest.java
@@ -1,6 +1,6 @@
 package org.chzz.market.common;
 
-import org.chzz.market.common.confing.QuerydslConfig;
+import org.chzz.market.common.config.QuerydslConfig;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
